Fix stupid error where I used a temporary hash function as sequence generator...

+++ core/src/main/java/org/nzbhydra/database/SearchResultRepository.java
@@ -3,7 +3,7 @@ package org.nzbhydra.database;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SearchResultRepository extends JpaRepository<SearchResultEntity, Integer>{
+public interface SearchResultRepository extends JpaRepository<SearchResultEntity, Long> {
 
     SearchResultEntity findByIndexerAndIndexerGuid(IndexerEntity indexer, String indexerGuid);
 }
